Update tut2.cpp
I am new to Open Source projects and I want to start it out from your repository. 
I am your fan and I want to say that you are doing a great job, so much content you have on YouTube.
Because of you I was able to make a Game. Please keep up the Good work.

I know I am asking too much but can you make a video on how we can use Allegro 5, SDL 2 and sfml in one software or if it is possible.
Thanks, a big fan of yours.